+++ optuna/storages/__init__.py
@@ -1,10 +1,10 @@
 from typing import Union  # NOQA
 
-from optuna.storages.base import BaseStorage  # NOQA
-from optuna.storages.cached_storage import _CachedStorage
-from optuna.storages.in_memory import InMemoryStorage
+from optuna.storages._base import BaseStorage  # NOQA
+from optuna.storages._cached_storage import _CachedStorage
+from optuna.storages._in_memory import InMemoryStorage
 from optuna.storages.rdb.storage import RDBStorage
-from optuna.storages.redis import RedisStorage
+from optuna.storages._redis import RedisStorage
 
 
 def get_storage(storage):
