+++ tests/storages_tests/test_storages.py
@@ -14,9 +14,9 @@ import optuna
 from optuna.distributions import CategoricalDistribution
 from optuna.distributions import LogUniformDistribution
 from optuna.distributions import UniformDistribution
-from optuna.storages.base import DEFAULT_STUDY_NAME_PREFIX
+from optuna.storages._base import DEFAULT_STUDY_NAME_PREFIX
 from optuna.storages import BaseStorage
-from optuna.storages.cached_storage import _CachedStorage
+from optuna.storages import _CachedStorage
 from optuna.storages import InMemoryStorage
 from optuna.storages import RDBStorage
 from optuna.storages import RedisStorage
